% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_to.R
\name{align_to}
\alias{align_to}
\title{Align a raster stack to a target grid (bilinear)}
\usage{
align_to(src, target)
}
\arguments{
\item{src}{A terra SpatRaster.}

\item{target}{A terra SpatRaster used as target geometry.}
}
\value{
A SpatRaster aligned to target.
}
\description{
Align a raster stack to a target grid (bilinear)
}
