% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_occ_list.R
\name{gbif_occ_list}
\alias{gbif_occ_list}
\title{Download GBIF occurrences for multiple species}
\usage{
gbif_occ_list(species_vec, country_iso, years, limit)
}
\arguments{
\item{species_vec}{Character vector of species scientific names.}

\item{country_iso}{ISO2 country code.}

\item{years}{Length-2 vector.}

\item{limit}{Max records per species.}
}
\value{
Named list of sf objects (or NULL entries).
}
\description{
Download GBIF occurrences for multiple species
}
