% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_occ_sf.R
\name{gbif_occ_sf}
\alias{gbif_occ_sf}
\title{Download GBIF occurrences for one species and return sf points}
\usage{
gbif_occ_sf(
  scientific_name,
  country_iso = "IT",
  years = c(2000, 2020),
  limit = 20000
)
}
\arguments{
\item{scientific_name}{Species scientific name.}

\item{country_iso}{ISO2 country code (e.g., "IT").}

\item{years}{Length-2 vector: c(start, end).}

\item{limit}{Max records returned by GBIF call.}
}
\value{
An sf object (EPSG:4326) or NULL if none found.
}
\description{
Download GBIF occurrences for one species and return sf points
}
