% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_suitability_cube.R
\name{merge_suitability_cube}
\alias{merge_suitability_cube}
\title{Merge AOA, DI and HV cubes into a single multi-attribute stars object}
\usage{
merge_suitability_cube(
  AOA_cube,
  DI_cube,
  HV_cube,
  species_vec,
  time_values = c("present", "future")
)
}
\arguments{
\item{AOA_cube}{stars cube with attribute "AOA".}

\item{DI_cube}{stars cube with attribute "DI".}

\item{HV_cube}{stars cube with attribute "HV".}

\item{species_vec}{Character vector of species names to set on the taxon dimension.}

\item{time_values}{Character vector for time dimension (default: c("present","future")).}
}
\value{
A multi-attribute stars object containing AOA, DI, HV.
}
\description{
Merge AOA, DI and HV cubes into a single multi-attribute stars object
}
