% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_template}
\alias{params_template}
\title{Default parameters (edit this list in your workflow)}
\usage{
params_template()
}
\value{
A named list of parameters.
}
\description{
This function returns the same `params` list used in the tutorial scripts.
Users should assign it and edit values, e.g.:
`params <- params_template(); params$country_name <- "Italy"`.
}
